Organize components using ddd
