+++ src/components/categories/CategoryBuilder.tsx
@@ -2,6 +2,7 @@ import { Plus, Trash2, X } from 'lucide-react';
 import { memo, useCallback, useEffect, useRef, useState } from 'react';
 import { toast } from 'sonner';
 
+import { IconPicker } from '@/components/shared/IconPicker';
 import {
   AlertDialog,
   AlertDialogAction,
@@ -20,8 +21,6 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@
 import { useAuth } from '@/contexts/AuthContext';
 import { supabase } from '@/lib/supabaseClient';
 import type { CategoryDefinition, FieldDefinition, FieldType } from '@/types/types';
-
-import { IconPicker } from './IconPicker';
 
 // --- Helpers ---
 const slugify = (text: string) => {
