+++ src/components/items/ComparisonModal.tsx
