+++ src/components/items/ItemDetailView.tsx
