+++ src/components/items/ItemDetailsContent.tsx
